+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - 2022-??-??
 ### Fixed
 
+- Nintendo Switch: Pressing "system" button has a delay of 200ms before being able
+  to press it again. This helps switching the Joystick port in Unijoysticle.
 - 8BitDo: Added SNES30 to the DB. It correctly identifies it as 8BitDo gamepad
 
 ## [3.0.0-rc0] - 2022-01-31